MPP: Support single platform android projects in new Android source set model
